[ci] Reorganize build.py code
[ghstack-poisoned]
