+++ .github/workflows/scripts/build.py
@@ -1,7 +1,7 @@
-#!/usr/bin/env python3
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
 
-import ci_common  # isort: skip, early initialization happens here
-
+# -- stdlib --
 import argparse
 import glob
 import os
@@ -9,175 +9,20 @@ import platform
 import sys
 from pathlib import Path
 
+# -- third party --
+# -- own --
 from ci_common import misc
-from ci_common.dep import download_dep
-from ci_common.misc import (banner, get_cache_home, is_manylinux2014,
-                            path_prepend)
-from ci_common.python import path_prepend, setup_python
+from ci_common.android import build_android, setup_android_ndk
+from ci_common.compiler import setup_clang, setup_msvc
+from ci_common.llvm import setup_llvm
+from ci_common.misc import banner, is_manylinux2014
+from ci_common.python import setup_python
 from ci_common.sccache import setup_sccache
-from ci_common.tinysh import Command, git, sh
+from ci_common.tinysh import Command, git
+from ci_common.vulkan import setup_vulkan
 
 
 # -- code --
-@banner('Setup Clang')
-def setup_clang(as_compiler=True) -> None:
-    '''
-    Setup Clang.
-    '''
-    u = platform.uname()
-    if u.system == 'Linux':
-        pass
-    elif (u.system, u.machine) == ('Windows', 'AMD64'):
-        out = get_cache_home() / 'clang-15-v2'
-        url = 'https://github.com/taichi-dev/taichi_assets/releases/download/llvm15/clang-15.0.0-win-complete.zip'
-        download_dep(url, out, force=True)
-        clang = str(out / 'bin' / 'clang++.exe').replace('\\', '\\\\')
-        os.environ['TAICHI_CMAKE_ARGS'] += f' -DCLANG_EXECUTABLE={clang}'
-
-        if as_compiler:
-            os.environ['TAICHI_CMAKE_ARGS'] += (
-                f' -DCMAKE_CXX_COMPILER={clang}'
-                f' -DCMAKE_C_COMPILER={clang}')
-    else:
-        # TODO: unify all
-        pass
-
-
-@banner('Setup MSVC')
-def setup_msvc() -> None:
-    assert platform.system() == 'Windows'
-    os.environ['TAICHI_USE_MSBUILD'] = '1'
-
-    url = 'https://aka.ms/vs/17/release/vs_BuildTools.exe'
-    out = Path(
-        r'C:\Program Files (x86)\Microsoft Visual Studio\2022\BuildTools')
-    download_dep(
-        url,
-        out,
-        args=[
-            '--passive',
-            '--wait',
-            '--norestart',
-            '--includeRecommended',
-            '--add',
-            'Microsoft.VisualStudio.Workload.VCTools',
-            # NOTE: We are using the custom built Clang++,
-            #       so components below are not necessary anymore.
-            # '--add',
-            # 'Microsoft.VisualStudio.Component.VC.Llvm.Clang',
-            # '--add',
-            # 'Microsoft.VisualStudio.ComponentGroup.NativeDesktop.Llvm.Clang',
-            # '--add',
-            # 'Microsoft.VisualStudio.Component.VC.Llvm.ClangToolset',
-        ])
-
-
-@banner('Setup LLVM')
-def setup_llvm() -> None:
-    '''
-    Download and install LLVM.
-    '''
-    u = platform.uname()
-    if u.system == 'Linux':
-        if 'AMDGPU_TEST' in os.environ:
-            out = get_cache_home() / 'llvm15-amdgpu'
-            url = 'https://github.com/GaleSeLee/assets/releases/download/v0.0.4/taichi-llvm-15.0.0-linux.zip'
-        elif is_manylinux2014():
-            # FIXME: prebuilt llvm15 on ubuntu didn't work on manylinux2014 image of centos. Once that's fixed, remove this hack.
-            out = get_cache_home() / 'llvm15-manylinux2014'
-            url = 'https://github.com/ailzhang/torchhub_example/releases/download/0.3/taichi-llvm-15-linux.zip'
-        else:
-            out = get_cache_home() / 'llvm15'
-            url = 'https://github.com/taichi-dev/taichi_assets/releases/download/llvm15/taichi-llvm-15-linux.zip'
-        download_dep(url, out, strip=1)
-    elif (u.system, u.machine) == ('Darwin', 'arm64'):
-        out = get_cache_home() / 'llvm15-m1'
-        url = 'https://github.com/taichi-dev/taichi_assets/releases/download/llvm15/taichi-llvm-15-m1.zip'
-        download_dep(url, out, strip=1)
-    elif (u.system, u.machine) == ('Darwin', 'x86_64'):
-        out = get_cache_home() / 'llvm15-mac'
-        url = 'https://github.com/taichi-dev/taichi_assets/releases/download/llvm15/llvm-15-mac10.15.zip'
-        download_dep(url, out, strip=1)
-    elif (u.system, u.machine) == ('Windows', 'AMD64'):
-        out = get_cache_home() / 'llvm15'
-        url = 'https://github.com/python3kgae/taichi_assets/releases/download/llvm15_vs2019_clang/taichi-llvm-15.0.0-msvc2019.zip'
-        download_dep(url, out, strip=0)
-    else:
-        raise RuntimeError(f'Unsupported platform: {u.system} {u.machine}')
-
-    # We should use LLVM toolchains shipped with OS.
-    # path_prepend('PATH', out / 'bin')
-    os.environ['LLVM_DIR'] = str(out)
-
-
-@banner('Setup Android NDK')
-def setup_android_ndk() -> None:
-    # TODO: Auto install
-    s = platform.system()
-    if s != 'Linux':
-        raise RuntimeError(
-            f'Android NDK is only supported on Linux, but the current platform is {s}.'
-        )
-
-    ndkroot = Path(
-        os.environ.get('ANDROID_NDK_ROOT', '/android-sdk/ndk-bundle'))
-    toolchain = ndkroot / 'build/cmake/android.toolchain.cmake'
-    if not toolchain.exists():
-        raise RuntimeError(
-            f'ANDROID_NDK_ROOT is set to {ndkroot}, but the path does not exist.'
-        )
-
-    p = ndkroot.resolve()
-    os.environ['ANDROID_NDK_ROOT'] = str(p)
-    cmake_args = (f' -DCMAKE_TOOLCHAIN_FILE={toolchain}'
-                  ' -DANDROID_NATIVE_API_LEVEL=29'
-                  ' -DANDROID_ABI=arm64-v8a')
-    os.environ['ANDROID_CMAKE_ARGS'] = cmake_args.strip()
-    os.environ['TAICHI_CMAKE_ARGS'] += cmake_args
-    path_prepend('PATH', p / 'toolchains/llvm/prebuilt/linux-x86_64/bin')
-
-
-@banner('Setup Vulkan 1.3.236.0')
-def setup_vulkan():
-    u = platform.uname()
-    if u.system == "Linux":
-        url = 'https://sdk.lunarg.com/sdk/download/1.3.236.0/linux/vulkansdk-linux-x86_64-1.3.236.0.tar.gz'
-        prefix = get_cache_home() / 'vulkan-1.3.236.0'
-        download_dep(url, prefix, strip=1)
-        sdk = prefix / 'x86_64'
-        os.environ['VULKAN_SDK'] = str(sdk)
-        path_prepend('PATH', sdk / "bin")
-        path_prepend('LD_LIBRARY_PATH', sdk / 'lib')
-        os.environ['VK_LAYER_PATH'] = str(sdk / 'etc' / 'vulkan' /
-                                          'explicit_layer.d')
-    # elif (u.system, u.machine) == ("Darwin", "arm64"):
-    # elif (u.system, u.machine) == ("Darwin", "x86_64"):
-    elif (u.system, u.machine) == ('Windows', 'AMD64'):
-        url = 'https://sdk.lunarg.com/sdk/download/1.3.236.0/windows/VulkanSDK-1.3.236.0-Installer.exe'
-        prefix = get_cache_home() / 'vulkan-1.3.236.0'
-        download_dep(
-            url,
-            prefix,
-            args=[
-                '--accept-licenses',
-                '--default-answer',
-                '--confirm-command',
-                '--root',
-                prefix,
-                'install',
-                'com.lunarg.vulkan.sdl2',
-                'com.lunarg.vulkan.glm',
-                'com.lunarg.vulkan.volk',
-                'com.lunarg.vulkan.vma',
-                # 'com.lunarg.vulkan.debug',
-            ])
-        os.environ['VULKAN_SDK'] = str(prefix)
-        os.environ['VK_SDK_PATH'] = str(prefix)
-        path_prepend('PATH', prefix / "Bin")
-    else:
-        return
-
-
 @banner('Build Taichi Wheel')
 def build_wheel(python: Command, pip: Command) -> None:
     '''
@@ -242,17 +87,6 @@ def action_wheel():
             f'Failed to produce exactly one wheel file: {distfiles}')
 
 
-@banner('Build Taichi Android C-API Shared Library')
-def build_android(python: Command, pip: Command) -> None:
-    '''
-    Build the Taichi Android C-API shared library
-    '''
-    pip.install('-r', 'requirements_dev.txt')
-    python('setup.py', 'clean')
-    python('setup.py', 'build_ext')
-    sh('aarch64-linux-android-strip', 'build/libtaichi_c_api.so')
-
-
 def action_android():
     sccache, python, pip = setup_basic_build_env()
     setup_android_ndk()
@@ -265,7 +99,7 @@ def action_android():
     sccache('-s')
 
 
-@banner('Add AOT Demo Related Environment Variables')
+@banner('Add AOT Related Environment Variables')
 def add_aot_env():
     os.environ['TAICHI_REPO_DIR'] = os.getcwd()
     for p in Path(os.getcwd()).glob('**/cmake-install/c_api'):
